Do not push doco if 2.7, only on 3.5

+++ deploy.sh
@@ -9,19 +9,21 @@ if [ -n "$GITHUB_API_KEY2" ]; then
     make clean
     make html
     make html
-    cd "$TRAVIS_BUILD_DIR"
-    rm -rf .git/
-    cd doc/out/html
-    git config --global user.email "travis"
-    git config --global user.name "travis"
-    touch .nojekyll
-    git init
-    git add .
-    echo "Committing"
-    git commit -m init
-    # Make sure to make the output quiet, or else the API token will leak!
-    # This works because the API key can replace your password.
-    echo "Pushing"
-    git push -f -q "https://${GITHUB_API_KEY2}@${GH_REF}" master:gh-pages  > /dev/null 2>&1 && echo "Pushed"
+    if [ "$PYTHON_VERSION" != "2.7" ]; then
+        cd "$TRAVIS_BUILD_DIR"
+        rm -rf .git/
+        cd doc/out/html
+        git config --global user.email "travis"
+        git config --global user.name "travis"
+        touch .nojekyll
+        git init
+        git add .
+        echo "Committing"
+        git commit -m init
+        # Make sure to make the output quiet, or else the API token will leak!
+        # This works because the API key can replace your password.
+        echo "Pushing"
+        git push -f -q "https://${GITHUB_API_KEY2}@${GH_REF}" master:gh-pages  > /dev/null 2>&1 && echo "Pushed"
+    fi
 fi
 echo "Deploy script ending"
